requirements.txt
Flask>=2.0
Werkzeug>=2.0
matplotlib>=3.0
reportlab>=3.0
Pillow>=9.0
openpyxl>=3.0
APScheduler>=3.8
pytest>=7.0
